rust: refactoring

+++ src/lib/wordart/wasm-gen/src/matrix.rs
@@ -1,14 +1,14 @@
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, PartialEq, Debug, Serialize)]
 pub struct Matrix {
-  pub a: f64,
-  pub b: f64,
-  pub c: f64,
-  pub d: f64,
-  pub e: f64,
-  pub f: f64,
+  pub a: f32,
+  pub b: f32,
+  pub c: f32,
+  pub d: f32,
+  pub e: f32,
+  pub f: f32,
 }
 
 pub enum MatrixTransformError {
@@ -18,16 +18,16 @@ pub enum MatrixTransformError {
 impl Matrix {
   pub fn new() -> Matrix {
     Matrix {
-      a: 1f64,
-      b: 0f64,
-      c: 0f64,
-      d: 1f64,
-      e: 0f64,
-      f: 0f64,
+      a: 1f32,
+      b: 0f32,
+      c: 0f32,
+      d: 1f32,
+      e: 0f32,
+      f: 0f32,
     }
   }
 
-  pub fn between(from: &Matrix, to: &Matrix, ratio: f64) -> Matrix {
+  pub fn between(from: &Matrix, to: &Matrix, ratio: f32) -> Matrix {
     Matrix {
       a: from.a + (to.a - from.a) * ratio,
       b: from.b + (to.b - from.b) * ratio,
@@ -38,29 +38,29 @@ impl Matrix {
     }
   }
 
-  pub fn translate_mut(&mut self, x: f64, y: f64) {
+  pub fn translate_mut(&mut self, x: f32, y: f32) {
     self.e += self.a * x + self.c * y;
     self.f += self.b * x + self.d * y;
   }
 
-  pub fn translate(mut self, x: f64, y: f64) -> Matrix {
+  pub fn translate(mut self, x: f32, y: f32) -> Matrix {
     self.translate_mut(x, y);
     self
   }
 
-  pub fn scale_mut(&mut self, x: f64, y: f64) {
+  pub fn scale_mut(&mut self, x: f32, y: f32) {
     self.a *= x;
     self.b *= x;
     self.c *= y;
     self.d *= y;
   }
 
-  pub fn scale(mut self, x: f64, y: f64) -> Matrix {
+  pub fn scale(mut self, x: f32, y: f32) -> Matrix {
     self.scale_mut(x, y);
     self
   }
 
-  pub fn rotate_mut(&mut self, angle: f64) {
+  pub fn rotate_mut(&mut self, angle: f32) {
     let cos = angle.cos();
     let sin = angle.sin();
     let Matrix { a, b, c, d, .. } = *self;
@@ -70,29 +70,29 @@ impl Matrix {
     self.d = d * cos - b * sin;
   }
 
-  pub fn rotate(mut self, angle: f64) -> Matrix {
+  pub fn rotate(mut self, angle: f32) -> Matrix {
     self.rotate_mut(angle);
     self
   }
 
-  pub fn skew_x_mut(&mut self, angle: f64) {
+  pub fn skew_x_mut(&mut self, angle: f32) {
     let tan = angle.tan();
     self.c += self.a * tan;
     self.d += self.b * tan;
   }
 
-  pub fn skew_x(mut self, angle: f64) -> Matrix {
+  pub fn skew_x(mut self, angle: f32) -> Matrix {
     self.skew_x_mut(angle);
     self
   }
 
-  pub fn skew_y_mut(&mut self, angle: f64) {
+  pub fn skew_y_mut(&mut self, angle: f32) {
     let tan = angle.tan();
     self.a += self.c * tan;
     self.b += self.d * tan;
   }
 
-  pub fn skew_y(mut self, angle: f64) -> Matrix {
+  pub fn skew_y(mut self, angle: f32) -> Matrix {
     self.skew_y_mut(angle);
     self
   }
@@ -101,7 +101,7 @@ impl Matrix {
     self.transform_values_mut(matrix.a, matrix.b, matrix.c, matrix.d, matrix.e, matrix.f);
   }
 
-  pub fn transform_values_mut(&mut self, pa: f64, pb: f64, pc: f64, pd: f64, pe: f64, pf: f64) {
+  pub fn transform_values_mut(&mut self, pa: f32, pb: f32, pc: f32, pd: f32, pe: f32, pf: f32) {
     let a = self.a;
     let b = self.b;
     let c = self.c;
@@ -121,12 +121,12 @@ impl Matrix {
 
   pub fn transform_values(
     mut self,
-    pa: f64,
-    pb: f64,
-    pc: f64,
-    pd: f64,
-    pe: f64,
-    pf: f64,
+    pa: f32,
+    pb: f32,
+    pc: f32,
+    pd: f32,
+    pe: f32,
+    pf: f32,
   ) -> Matrix {
     self.transform_values_mut(pa, pb, pc, pd, pe, pf);
     self
@@ -136,7 +136,7 @@ impl Matrix {
     self.set_mut(1.0, 0.0, 0.0, 1.0, 0.0, 0.0);
   }
 
-  pub fn set_mut(&mut self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64) {
+  pub fn set_mut(&mut self, a: f32, b: f32, c: f32, d: f32, e: f32, f: f32) {
     self.a = a;
     self.b = b;
     self.c = c;
@@ -145,16 +145,16 @@ impl Matrix {
     self.f = f;
   }
 
-  pub fn det(&self) -> f64 {
+  pub fn det(&self) -> f32 {
     self.a * self.d - self.b * self.c
   }
 
   pub fn inverse_mut(&mut self) -> Result<(), MatrixTransformError> {
     let mut det = self.det();
-    if det == 0f64 {
+    if det == 0f32 {
       return Err(MatrixTransformError::DeterminantIsZero);
     }
-    det = 1f64 / det;
+    det = 1f32 / det;
     let a = self.d * det;
     let b = -self.b * det;
     let c = -self.c * det;
